@@ -1,17 +1,9 @@
-// The Vue build version to load with the `import` command
-// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
-import Vue from 'vue';
-import App from './App';
-import router from './router';
-import VueScrollTo from 'vue-scrollto';
+import { createApp } from "vue";
+import App from "./App.vue";
+import router from "./router";
+import VueScrollTo from "vue-scrollto";
 
-Vue.config.productionTip = false;
-Vue.use(VueScrollTo);
-
-/* eslint-disable no-new */
-new Vue({
-  el: '#app',
-  router,
-  components: { App },
-  template: '<App/>'
-});
+createApp(App)
+  .use(VueScrollTo)
+  .use(router)
+  .mount("#app");
