@@ -1,33 +1,65 @@
 <template>
-  <div id="app">
+  <div>
     <header>
       <nav>
         <div class="nav-wrapper">
-          <h1><a href="/#/" title="Recruitment Ipsum">Recruitment Ipsum</a></h1>
+          <h1>
+            <router-link :to="{ name: 'Home' }" title="Recruitment Ipsum"
+              >Recruitment Ipsum</router-link
+            >
+          </h1>
           <ul>
-            <li><a href="/#/about" title="About Recruitment Ipsum">About <span class="long-about">Recruitment Ipsum</span></a></li>
+            <li>
+              <router-link
+                :to="{ name: 'About' }"
+                title="About Recruitment Ipsum"
+                >About
+                <span class="long-about">Recruitment Ipsum</span></router-link
+              >
+            </li>
           </ul>
         </div>
       </nav>
     </header>
-    <router-view/>
+    <router-view />
     <footer>
       <p>
-        Created by <a href="https://twitter.com/jeroenooms" target="_blank"><span>Jeroen Ooms</span> <i class="fab fa-twitter"></i></a> with 
-        <a href="https://vuejs.org" target="_blank"><span>Vue.js</span> <i class="fab fa-vuejs"></i></a>,
-        <a href="https://fontawesome.com/" target="_blank"><span>Font Awesome</span> <i class="fab fa-font-awesome-flag"></i></a> and
-        <a href="https://pexels.com/" target="_blank"><span>Pexels</span> <i class="fas fa-image"></i></a>
+        Created by
+        <a href="https://twitter.com/jeroenooms" target="_blank"
+          ><span>Jeroen Ooms</span> <i class="fab fa-twitter"></i
+        ></a>
+        with
+        <a href="https://vuejs.org" target="_blank"
+          ><span>Vue.js</span> <i class="fab fa-vuejs"></i></a
+        >,
+        <a href="https://fontawesome.com/" target="_blank"
+          ><span>Font Awesome</span> <i class="fab fa-font-awesome-flag"></i
+        ></a>
+        and
+        <a href="https://pexels.com/" target="_blank"
+          ><span>Pexels</span> <i class="fas fa-image"></i
+        ></a>
       </p>
-      <p>View the source code on <a href="https://github.com/jeroenoomsNL/recruitment-ipsum" target="_blank"><span>Github</span> <i class="fab fa-github"></i></a></p>
-      <p>Special thanks to all those creative recruiters out there! <i class="fas fa-heart"></i></p>
+      <p>
+        View the source code on
+        <a
+          href="https://github.com/jeroenoomsNL/recruitment-ipsum"
+          target="_blank"
+          ><span>Github</span> <i class="fab fa-github"></i
+        ></a>
+      </p>
+      <p>
+        Special thanks to all those creative recruiters out there!
+        <i class="fas fa-heart"></i>
+      </p>
     </footer>
   </div>
 </template>
 
 <script>
 export default {
-  name: 'App'
-}
+  name: "App",
+};
 </script>
 
 <style lang="scss">
@@ -36,7 +68,7 @@ body {
   color: #173e43;
   margin: 0;
   line-height: 1.4;
-  font-family: 'Open Sans', sans-serif;
+  font-family: "Open Sans", sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
@@ -46,7 +78,7 @@ body {
 }
 
 *:focus {
-    outline: none;
+  outline: none;
 }
 
 a {
@@ -60,7 +92,7 @@ h2 {
 }
 
 blockquote {
-  font-family: Georgia,Source Serif Pro,serif;
+  font-family: Georgia, Source Serif Pro, serif;
   font-style: italic;
   font-size: 1.2em;
   max-width: 70%;
@@ -68,97 +100,95 @@ blockquote {
   text-align: center;
 }
 
-#app {
+header {
+  background-color: #173e43;
 
-  header {
-    background-color: #173e43; 
-
-    @media (min-width: 900px) {
-      background: url('./assets/background.jpg');
-      background-size: cover;
-      height: 500px;
-    }
+  @media (min-width: 900px) {
+    background: url("./assets/background.jpg");
+    background-size: cover;
+    height: 500px;
   }
+}
 
-  nav {
-    background-color: #3fb0ac;
-    border-bottom: 1px solid darken(#3fb0ac,8);
+nav {
+  background-color: #3fb0ac;
+  border-bottom: 1px solid darken(#3fb0ac, 8);
 
-    .nav-wrapper {
-      align-items: center;
-      display: flex;
-      justify-content: space-between;
-      margin: 0 auto;
-      padding: 20px;
-      max-width: 900px;
-    }
-
-    h1, ul {
-      color: #f1f1f1;
-      display: inline-block;
-      margin: 0;
-
-      a {
-        color: white;
-        text-decoration: none;
-        white-space: nowrap;
-      }
-    }
-
-    ul {
-      list-style: none;
-      padding: 0;
-
-      .long-about {
-        display: none;
-
-        @media (min-width: 900px) {
-          display: inline;
-        }
-      }
-    }
-  }
-
-  .page {
-    @media (min-width: 900px) {
-      margin-top: -250px;
-    }
-  }
-
-  .container {
-    background-color: white;
-    margin-left: auto;
-    margin-right: auto;
-    max-width: 900px;
-    padding: 20px;
-    position: relative;
-
-    @media (min-width: 900px) {
-      padding: 50px;
-    }
-  }
-
-  footer {
-    background: #173e43;
-    color: white;
+  .nav-wrapper {
+    align-items: center;
+    display: flex;
+    justify-content: space-between;
     margin: 0 auto;
     padding: 20px;
-    text-align: center;
     max-width: 900px;
+  }
 
-    @media (min-width: 900px) {
-      margin-top: 30px;
-      padding: 50px;
-    }
+  h1,
+  ul {
+    color: #f1f1f1;
+    display: inline-block;
+    margin: 0;
 
     a {
       color: white;
       text-decoration: none;
       white-space: nowrap;
+    }
+  }
 
-      span {
-        text-decoration: underline;
+  ul {
+    list-style: none;
+    padding: 0;
+
+    .long-about {
+      display: none;
+
+      @media (min-width: 900px) {
+        display: inline;
       }
+    }
+  }
+}
+
+.page {
+  @media (min-width: 900px) {
+    margin-top: -250px;
+  }
+}
+
+.container {
+  background-color: white;
+  margin-left: auto;
+  margin-right: auto;
+  max-width: 900px;
+  padding: 20px;
+  position: relative;
+
+  @media (min-width: 900px) {
+    padding: 50px;
+  }
+}
+
+footer {
+  background: #173e43;
+  color: white;
+  margin: 0 auto;
+  padding: 20px;
+  text-align: center;
+  max-width: 900px;
+
+  @media (min-width: 900px) {
+    margin-top: 30px;
+    padding: 50px;
+  }
+
+  a {
+    color: white;
+    text-decoration: none;
+    white-space: nowrap;
+
+    span {
+      text-decoration: underline;
     }
   }
 }
